+++ presentation/src/main/java/com/dietician/presentation/model/Food.kt
@@ -4,11 +4,11 @@ data class Food (
     val id: Long,
     val name: String,
     val fat: Double,
-    val protine: Long,
-    val carbohydrate: Long,
+    val protine: Double,
+    val carbohydrate: Double,
     val isVeg: Boolean,
     val type: Int,
     val foodCategory: Int,
-    val foodQuantity: Long
+    val foodQuantity: Double
 
 )
